% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_interaction.R
\name{add.int}
\alias{add.int}
\title{Function to add interactions}
\usage{
add.int(d.train)
}
\arguments{
\item{d.train}{a data frame with feature columns and gold standard column}
}
\description{
\code{add.int} adds interaction features to a data frame
}
\examples{
add.int(d.train)
}

