% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aupr.R
\name{calc.aupr}
\alias{calc.aupr}
\title{Calculate AUPR and AUROC}
\usage{
calc.aupr(pred, gs)
}
\arguments{
\item{pred}{a numeric vector of predictions}

\item{gs}{a logical vector of gold standards specifying 
whether the predictions are in the gold standard}
}
\description{
\code{calc.aupr} takes a list of predictions and gold standard 
and returns AUPR, AUROC, precision at 10% recall and 
sensitivity at 1% false positive rate.
}

