% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_similarity.R
\name{chisq.perm}
\alias{chisq.perm}
\title{Chi-square permutation test}
\usage{
chisq.perm(df1, df2, size = 10000, nperm = 10000)
}
\arguments{
\item{df1}{a one-column data frame storing feature values}

\item{df2}{a one-column data frame storing feature values}

\item{size}{an integer specifying sample size; default size to 10000}

\item{nperm}{an integer specifying number of permutations; default to 10000}
}
\description{
\code{chisq.perm} function tests for categorical feature similarity and returns an empirical p-value of significance.
}
\examples{
p.list <- chisq.perm(df1,df2)
}

