% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_camera_cast_object.R
\encoding{UTF-8}
\name{create_camera_cast_object}
\alias{create_camera_cast_object}
\title{Create a camera cast object}
\usage{
create_camera_cast_object(skeleton_list, sensorId)
}
\arguments{
\item{skeleton_list}{list of lists where every list element can be coerced to
a skeleton object}

\item{sensorId}{character string of sensorId}
}
\description{
create camera cast object, which is a list of skeletons of one camera
(for internal use mostly)
}
\author{
Janek
}

