% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_skeleton_object.R
\encoding{UTF-8}
\name{create_skeleton_object}
\alias{create_skeleton_object}
\title{Create a skeleton Object}
\usage{
create_skeleton_object(skeleton)
}
\arguments{
\item{skeleton}{list of elements to create a skeleton object}
}
\description{
create skeleton object, which is one picture of one camera (for internal use
mostly)
}
\author{
Janek
}

