% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_similarity.R
\name{feature.similarity}
\alias{feature.similarity}
\title{Wrapper function to assess feature similarity.}
\usage{
feature.similarity(d1, d2, verbose = TRUE)
}
\arguments{
\item{d1}{data frame 1}

\item{d2}{data frame 2}

\item{verbose}{logical, optional parameter}
}
\value{
a data frame of p-values
}
\description{
\code{feature.similarity} is a wrapper function to test for similarities between feature 
columns in two data frames d1 and d2.
}
\examples{
feature.similarity(d1,d2)
}

