% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fix_timestamps.R
\encoding{UTF-8}
\name{fix_timestamps}
\alias{fix_timestamps}
\title{Fix missing timestamps}
\usage{
fix_timestamps(ensemble)
}
\arguments{
\item{ensemble}{A camera cast ensemble object, created by read_motion_data
(or create_camera_cast_ensemble)}
}
\description{
Fix missing timestamps in situations where a camera didn't recognice a person
}
\author{
Janek
}

