% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get.all.data}
\alias{get.all.data}
\title{Function to load data}
\usage{
get.all.data(path, chrom, tf, ct, n, params, seqlen = 24, coords = FALSE,
  chip = FALSE)
}
\arguments{
\item{path}{a directory specifying the path to all calculated scores}

\item{chrom}{a character string, e.g. "chr12"}

\item{tf}{a character string, e.g. "CTCF"}

\item{ct}{a character string, e.g. "Hepg2"}

\item{n}{a numeric value}

\item{params}{a list of data frames returned by \code{em}}

\item{seqlen}{an integer specifying sequencing length, options are 24, 36, 40, 50, 75 and 100}

\item{coords}{logical}

\item{chip}{logical}
}
\description{
\code{get.all.data} loads data for sparse logistic regression training and making predictions.
provides sampling of loaded data as an option.
}

