% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get.cut.count}
\alias{get.cut.count}
\title{function to get motif-associated accessibility cut count}
\usage{
get.cut.count(path, chrom, tf, ct, n, seqlen = 24)
}
\arguments{
\item{path}{a directory path}

\item{chrom}{a character string, e.g. "chr12"}

\item{tf}{a character string, e.g. "CTCF"}

\item{ct}{a character string, e.g. "Hepg2"}

\item{n}{a numeric value}

\item{seqlen}{an integer, 24, 36, 50 or 100}
}
\description{
\code{get.cut.count}
}

