% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/joint_numbers.R
\docType{data}
\name{joint_numbers}
\alias{joint_numbers}
\title{Joint Numbers}
\format{A data frame with 25 rows and 2 variables}
\description{
Joint numbers with their corresponding part of the body
}
\keyword{datasets}

