% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kinematic_model.R
\docType{data}
\name{kinematic_model}
\alias{kinematic_model}
\title{kinematic_model}
\format{A matrix with 25 rows and columns}
\description{
Contains a 25x25 neighbourhood matrix. 1 for connected joints, 0 for not connected
}
\keyword{datasets}

