% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_similarity.R
\name{ks.perm}
\alias{ks.perm}
\title{KS permutation test}
\usage{
ks.perm(df1, df2, size = 500, nperm = 10000)
}
\arguments{
\item{df1}{a one-column data frame storing feature values}

\item{df2}{a one-column data frame storing feature values}

\item{size}{an integer specifying sample size; default size to 500}

\item{nperm}{an integer specifying number of permutations; default nperm to 10000}
}
\description{
\code{ks.perm} function tests for similarity between continous features and returns an empirical p-value of significance.
}
\examples{
p.list <- ks.perm(df1,df2)
}

