% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocap_em.R
\name{loglikfun}
\alias{loglikfun}
\title{Internal function}
\usage{
loglikfun(count, mu1, theta1, mu2, theta2, prob0, prob1, zi0, zi1)
}
\description{
log-likelihood function; setting 0s to .Machine$double.xmin to avoid numerical underflow of log(0)
}

