% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMotionSingleFrame.R
\name{plotMotionSingleFrame}
\alias{plotMotionSingleFrame}
\title{Plot a single frame of XYZ location}
\usage{
plotMotionSingleFrame(xyz, childs, frame, lims, rotateAngle = 0,
  viewAngle = 40, mainTitle = frame, ...)
}
\arguments{
\item{xyz}{desc}

\item{childs}{desc}

\item{frame}{desc}

\item{lims}{desc}

\item{rotateAngle}{desc}

\item{viewAngle}{desc}

\item{mainTitle}{desc}

\item{...}{desc}
}
\description{
This function allows you to...
}
\examples{
asfFilePath <- system.file("extdata", "lambada.asf", package = "mocap")
asf <- readASF(asfFilePath)
amcFilePath <- system.file("extdata", "lambada.amc", package = "mocap")
amc <- readAMC(amcFilePath, asf)
xyz <- getMotionData(asf, amc)
limX <- c(min(unlist(lapply(xyz, function(x) min(x[, 1])))),
  max(unlist(lapply(xyz, function(x) max(x[, 1])))))
limY <- c(min(unlist(lapply(xyz, function(x) min(x[, 2])))),
  max(unlist(lapply(xyz, function(x) max(x[, 2])))))
limZ <- c(min(unlist(lapply(xyz, function(x) min(x[, 3])))),
  max(unlist(lapply(xyz, function(x) max(x[, 3])))))
lims <- list(limX = limX, limY = limY, limZ = limZ)
plotMotionSingleFrame(xyz, amc$childs, 1)
}
