% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.camera_cast.R
\encoding{UTF-8}
\name{print.camera_cast}
\alias{print.camera_cast}
\title{print camera cast}
\usage{
\method{print}{camera_cast}(x, ...)
}
\arguments{
\item{x}{camera_cast object}

\item{...}{further arguments passed to print}
}
\description{
print camera cast objects
}
\author{
Janek
}

