% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.camera_cast_ensemble.R
\encoding{UTF-8}
\name{print.camera_cast_ensemble}
\alias{print.camera_cast_ensemble}
\title{print camera cast ensemble}
\usage{
\method{print}{camera_cast_ensemble}(object)
}
\description{
print camera cast ensemble
}
\author{
Janek
}

