% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_motion_data.R
\encoding{UTF-8}
\name{read_motion_data}
\alias{read_motion_data}
\title{Read in Motion Capturing Data}
\usage{
read_motion_data(file)
}
\arguments{
\item{file}{Path to JSON file}
}
\value{
Returns a camera cast ensemble object
}
\description{
Read Motion Capturing Data from Json file
}
\author{
Janek
}

