% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAMC.R
\name{readAMC}
\alias{readAMC}
\title{Read and Parse a AMC file}
\usage{
readAMC(amcFilePath, asf, in2cm = TRUE)
}
\arguments{
\item{amcFilePath}{path to a .asf file}

\item{asf}{ASF object read with \code{readASF()}}

\item{in2cm}{desc}
}
\description{
This function allows you to...
}
\examples{
asfFilePath <- system.file("extdata", "lambada.asf", package = "mocap")
asf <- readASF(asfFilePath)
amcFilePath <- system.file("extdata", "lambada.amc", package = "mocap")
amc <- readAMC(amcFilePath, asf)
}
