% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readASF.R
\name{readASF}
\alias{readASF}
\title{Read and Parse a AFS file}
\usage{
readASF(asfFilePath, in2cm = TRUE)
}
\arguments{
\item{asfFilePath}{path to a .asf file}

\item{in2cm}{desc}
}
\description{
This function allows you to...
}
\examples{
asfFilePath <- system.file("extdata", "lambada.asf", package = "mocap")
asf <- readASF(asfFilePath)
}
