% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mocap.R
\name{run.Mocap}
\alias{run.Mocap}
\title{generate motif/genomic region-assoicated accessibility scores}
\usage{
run.Mocap(bedCount, params)
}
\arguments{
\item{bedCount}{a n x 4 data frame, where columns specify chr, start, end, count}

\item{params}{a b x 6 data frame containing \code{em} output parameter set}
}
\value{
a list containing an integer cutOff value, 
a numeric vector of binary accessibility classifications and
a numeric vector of accessibility scores
}
\description{
\code{run.Mocap} takes as input a BED-4 formatted data frame and \code{em}
returned parameter set and output a cutoff value, a vector of binary classifications of 
genomic (motif) regions and a vector of log-likelihood accessibility scores.
}

