% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbisquare.R
\name{sample.mapping}
\alias{sample.mapping}
\title{cross-sample mapping function}
\usage{
sample.mapping(xdir, ydir, gamma, ind.test, cores, exp = "ATAC-Seq",
  verbose = TRUE)
}
\arguments{
\item{xdir}{a path to directory storing the x files}

\item{ydir}{a path to directory storing the y files}

\item{gamma}{a numeric value}

\item{ind.test}{an integer}

\item{cores}{an integer}

\item{exp}{a character string, either "ATAC-Seq" or "DNase-Seq"}

\item{verbose}{logical}
}
\description{
\code{sample.mapping} makes cross-sample mappings between models and samples.
}
\examples{
sample.mapping(xdir,ydir,gamma,ind.test,cores)
}

