% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/select_joint.R
\encoding{UTF-8}
\name{select_joint}
\alias{select_joint}
\title{Select one joint from a camera cast}
\usage{
select_joint(cast, joint, include_error = FALSE)
}
\arguments{
\item{cast}{camera cast object}

\item{joint}{numeric. Integer of the joint number to extract}

\item{include_error}{logical. Should the kinect specific error value be included}
}
\description{
Select one joint from a camera cast object.
}
\author{
Janek
}

