% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_mat.R
\name{select_mat}
\alias{select_mat}
\title{select matrices}
\usage{
select_mat(xdir, ydir, gamma, ind.test)
}
\arguments{
\item{xdir}{a directory containing pair-wise feature similarity files}

\item{ydir}{a directory containing cross-prediction performance files}

\item{gamma}{a numeric value specifying a hyperparameter}

\item{ind.test}{a vector of integers specifying hold-out data}
}
\description{
\code{select_mat} constructs  matrices for weighted least squares regression.
}

