% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/select_sensor.R
\encoding{UTF-8}
\name{select_sensor}
\alias{select_sensor}
\title{Select sensor from cast esemble}
\usage{
select_sensor(ensemble, sensor)
}
\arguments{
\item{ensemble}{camera_cast_ensemble object}

\item{sensor}{Integer value or sensorId}
}
\description{
Select one sensor from a camera cast esemble
}
\author{
Janek
}

