% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocap_em.R
\name{sim.count}
\alias{sim.count}
\title{simulate count data}
\usage{
sim.count(num, prob0.true, prob1.true, mu1.true, theta1.true, mu2.true,
  theta2.true)
}
\arguments{
\item{num}{an integer specifying the total number of counts}

\item{prob0.true}{a numeric value < 0.3}

\item{prob1.true}{a numeric value > 0.8}

\item{mu1.true}{a numeric value}

\item{theta1.true}{a numeric value}

\item{mu2.true}{a numeric value, where mu1.true << mu2.true}

\item{theta2.true}{a numeric value}
}
\value{
a numeric vector of simulated cut counts
}
\description{
\code{sim.count} simulate cut count data from a mixture of zero-inflated negative 
binomial distributions
}

