% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotionData.R
\name{getMotionData}
\alias{getMotionData}
\title{Get XYZ motion data from a pair of ASF/AMC files}
\usage{
getMotionData(asf, amc, verbose = FALSE)
}
\arguments{
\item{asf}{ASF object read with \code{readASF()}}

\item{amc}{AMC object read with \code{readAMC()}}

\item{verbose}{desc}
}
\description{
This function allows you to...
}
\examples{
asfFilePath <- system.file("extdata", "lambada.asf", package = "mocap")
asf <- readASF(asfFilePath)
amcFilePath <- system.file("extdata", "lambada.amc", package = "mocap")
amc <- readAMC(amcFilePath, asf)
xyz <- getMotionData(asf, amc)
}
