% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMotionMovie.R
\name{makeMotionMovie}
\alias{makeMotionMovie}
\title{Make a movie out of motion data}
\usage{
makeMotionMovie(asf, amc, xyz, type = "gif", framesPerSecond = 120,
  skipNFrames = 1, fileName = paste0("motion.", type), imgName = "motion",
  rotateMotion = FALSE, viewAngle = 40, ...)
}
\arguments{
\item{asf}{desc}

\item{amc}{desc}

\item{xyz}{desc}

\item{type}{desc}

\item{framesPerSecond}{desc}

\item{skipNFrames}{desc}

\item{fileName}{desc}

\item{imgName}{desc}

\item{rotateMotion}{desc}

\item{viewAngle}{desc}

\item{...}{desc}
}
\description{
This function allows you to...
}
\examples{
asfFilePath <- system.file("extdata", "lambada.asf", package = "mocap")
asf <- readASF(asfFilePath)
amcFilePath <- system.file("extdata", "lambada.amc", package = "mocap")
amc <- readAMC(amcFilePath, asf)
xyz <- getXYZLocationData(asf, amc)
makeMotionMovie(asf, amc, xyz, skipNFrames = 4)
}
