% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.camera_cast_ensemble.R
\encoding{UTF-8}
\name{print.camera_cast_ensemble}
\alias{print.camera_cast_ensemble}
\title{print camera cast ensemble}
\usage{
\method{print}{camera_cast_ensemble}(x, ...)
}
\arguments{
\item{x}{camera_cast_ensemble object}

\item{...}{further arguments passed to print}
}
\description{
print camera cast ensemble objects
}
\author{
Janek
}

