% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAMC.R
\name{readAMC}
\alias{readAMC}
\title{Read and Parse a AMC file}
\usage{
readAMC(amcFilePath, asf, in2cm = TRUE)
}
\arguments{
\item{amcFilePath}{path to a .asf file}

\item{asf}{ASF object read with \code{readASF()}}

\item{in2cm}{a boolean indicating whether to convert inches to centimeters, defaults to TRUE}
}
\value{
a list containing:
\item{D}{a matrix of dimensions bones x frames containing raw (or converted
to cm) AMC values}
\item{degrees}{a named list containing for each bone the nFrames x 3 matrix
with AMC Euler Angles}
\item{nFrames}{no. of frames found in the AMC file}
\item{skeleton}{a data frame containing the details for each bone, e.g. DOF,
length, direction, angle}
}
\description{
This function reads in and parses a AMC Motion Capture file and returns
several useful objects
}
\note{
Only ASF/AMC files from the CMU Graphics Lab Motion Capture Database have been tested.
}
\examples{
asfFilePath <- system.file("extdata", "lambada.asf", package = "mocap")
asf <- readASF(asfFilePath)
amcFilePath <- system.file("extdata", "lambada.amc", package = "mocap")
amc <- readAMC(amcFilePath, asf)

}
\references{
\url{http://giorasimchoni.com/}
\url{http://mocap.cs.cmu.edu/}
}
