% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readASF.R
\name{readASF}
\alias{readASF}
\title{Read and Parse a AFS file}
\usage{
readASF(asfFilePath, in2cm = TRUE)
}
\arguments{
\item{asfFilePath}{path to a .asf file}

\item{in2cm}{a boolean indicating whether to convert inches to centimeters,
defaults to TRUE}
}
\value{
a list containing:
\item{skeleton}{a data frame containing the details for each bone, e.g. DOF,
length, direction, angle - DO NOT USE, use the one returned from \code{readAMC}}
\item{childs}{a named list containing a vector of "children" bones for each
"parent" bone}
\item{CMatList}{a named list with the C and Cinv angle-axis matrices for each bone}
\item{len}{the length parameter in the ASF file, scaling factor all lengths,
in CMU files always 0.45}
\item{angleType}{the angle parameter in the ASF file, "deg" or "rad",
in CMU files always "deg"}
}
\description{
This function reads in and parses a AFS Motion Capture file and returns
several useful objects
}
\note{
Only ASF/AMC files from the CMU Graphics Lab Motion Capture Database have been tested.
}
\examples{
asfFilePath <- system.file("extdata", "lambada.asf", package = "mocap")
asf <- readASF(asfFilePath)

}
\references{
\url{http://giorasimchoni.com/}
\url{http://mocap.cs.cmu.edu/}
}
